workspaces/rabbit-zed/pyproject.toml
[project]
name = "rabbit-zed"
version = "0.0.1"
requires-python = ">=3.12,<3.13"
dependencies = ["nats-py", "cython", "numpy", "opencv-python", "pyopengl", "pyzed"]
